backend/requirements.txt
Django==3.2.7
djangorestframework==3.11.2
psycopg2-binary==2.8.5
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
faker==8.12.1
Pillow==7.1.2
